requirements.txt
amqp==5.1.1
asgiref==3.3.4
attrs==22.1.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
cffi==1.15.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.3
django-celery-results==2.2.0
django-import-export==2.5.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Django-Verify-Email==1.0.7
djangorestframework==3.13.1
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.8.3
incremental==21.3.0
kombu==5.1.0
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.1.5
Pillow==8.2.0
prompt-toolkit==3.0.31
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
Twisted==22.4.0
txaio==22.2.1
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.6.0
zope.interface==5.4.0
